refactor, removed api and moved volumes_fix_permissions
