+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/client_api.py
@@ -218,18 +218,6 @@ class DynamicSidecarClient:
             )
 
     @log_decorator(logger=logger)
-    async def service_volumes_fix_permissions(
-        self, dynamic_sidecar_endpoint: str
-    ) -> None:
-        url = get_url(dynamic_sidecar_endpoint, "/v1/containers/volumes/permissions:fix")
-
-        response = await self._client.post(url)
-        if response.status_code != status.HTTP_204_NO_CONTENT:
-            raise DynamicSidecarUnexpectedResponseStatus(
-                response, "volumes fix permissions"
-            )
-
-    @log_decorator(logger=logger)
     async def service_pull_output_ports(
         self, dynamic_sidecar_endpoint: str, port_keys: Optional[List[str]] = None
     ) -> int:
