+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/events.py
@@ -286,15 +286,8 @@ class PrepareServicesEnvironment(DynamicSchedulerEvent):
             logger.debug(
                 "Creating dirs from service outputs labels: %s", service_outputs_labels
             )
-            await logged_gather(
-                *[
-                    dynamic_sidecar_client.service_outputs_create_dirs(
-                        dynamic_sidecar_endpoint, service_outputs_labels
-                    ),
-                    dynamic_sidecar_client.service_volumes_fix_permissions(
-                        dynamic_sidecar_endpoint
-                    ),
-                ]
+            await dynamic_sidecar_client.service_outputs_create_dirs(
+                dynamic_sidecar_endpoint, service_outputs_labels
             )
 
             scheduler_data.dynamic_sidecar.service_environment_prepared = True
