+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/core/utils.py
@@ -23,6 +23,8 @@ from tenacity.before_sleep import before_sleep_log
 from tenacity.stop import stop_after_attempt
 from tenacity.wait import wait_fixed
 
+from ..modules.mounted_fs import MountedVolumes, get_mounted_volumes
+
 CommandResult = namedtuple("CommandResult", "finished_without_errors, decoded_stdout")
 
 TEMPLATE_SEARCH_PATTERN = r"%%(.*?)%%"
@@ -156,3 +158,20 @@ def assemble_container_names(validated_compose_content: str) -> List[str]:
         service_data["container_name"]
         for service_data in parsed_compose_spec["services"].values()
     ]
+
+
+async def volumes_fix_permissions() -> None:
+    # NOTE: by creating a hidden file on all mounted volumes
+    # the same permissions are ensured and avoids
+    # issues when starting the services
+    mounted_volumes: MountedVolumes = get_mounted_volumes()
+    for volume_path in [
+        mounted_volumes.disk_inputs_path,
+        mounted_volumes.disk_outputs_path,
+    ] + list(mounted_volumes.disk_state_paths()):
+        hidden_file = volume_path / ".hidden_do_not_remove"
+        hidden_file.write_text(
+            f"Directory must not be empty.\nCreated by {__file__}.\nRequired by "
+            "oSPARC internals to properly enforce permissions on this "
+            "directory and all its files"
+        )
