+++ services/dynamic-sidecar/tests/unit/test_api_containers.py
@@ -507,14 +507,6 @@ async def test_container_create_outputs_dirs(
     assert mock_dir_watcher_on_any_event.call_count == 2 * len(mock_outputs_labels)
 
 
-async def test_volumes_fix_permissions(
-    test_client: TestClient, mock_dir_watcher_on_any_event: AsyncMock
-) -> None:
-    response = await test_client.post(f"/{API_VTAG}/containers/volumes/permissions:fix")
-    assert response.status_code == status.HTTP_204_NO_CONTENT, response.text
-    assert response.text == ""
-
-
 async def test_container_pull_output_ports(
     test_client: TestClient, mock_port_keys: List[str], mock_nodeports: None
 ) -> None:
